+++ tests/test_types.py
@@ -5,7 +5,7 @@ from datetime import date, datetime, time, timedelta
 from decimal import Decimal
 from enum import Enum, IntEnum
 from pathlib import Path
-from typing import List, NewType, Set
+from typing import List, NewType, Pattern, Set
 from uuid import UUID
 
 import pytest
@@ -1172,3 +1172,26 @@ def test_json_not_str():
         'msg': 'JSON object must be str, bytes or bytearray',
         'type': 'type_error.json',
     }
+
+
+def test_pattern():
+    class Foobar(BaseModel):
+        pattern: Pattern
+
+    f = Foobar(pattern=r'^whatev.r\d$')
+    # SRE_Pattern for 3.6, Pattern for 3.7
+    assert f.pattern.__class__.__name__ in {'SRE_Pattern', 'Pattern'}
+    # check it's really a proper pattern
+    assert f.pattern.match('whatever1')
+    assert not f.pattern.match(' whatever1')
+
+
+def test_pattern_error():
+    class Foobar(BaseModel):
+        pattern: Pattern
+
+    with pytest.raises(ValidationError) as exc_info:
+        Foobar(pattern=f'[xx')
+    assert exc_info.value.errors() == [
+        {'loc': ('pattern',), 'msg': 'Invalid regular expression', 'type': 'value_error.regex_pattern'}
+    ]