+++ HISTORY.rst
@@ -5,7 +5,8 @@ History
 
 v0.x.x (xxx-xx-xx)
 ..................
-* fix schema for ``timedelta`` as number, by @tiangolo
+* fix schema for ``timedelta`` as number, #325 by @tiangolo
+* prevent validators being called repeatedly after inheritance, #327 by @samuelcolvin
 
 v0.16.1 (2018-12-10)
 ....................