+++ HISTORY.rst
@@ -3,11 +3,12 @@
 History
 -------
 
-v0.x.x (xxx-xx-xx)
-..................
+v0.17.0 (unreleased)
+....................
 * fix schema for ``timedelta`` as number, #325 by @tiangolo
 * prevent validators being called repeatedly after inheritance, #327 by @samuelcolvin
 * prevent duplicate validator check in ipython, fix #312 by @samuelcolvin
+* add "Using Pydantic" section to docs, #323 by @tiangolo & #326 by @samuelcolvin
 
 v0.16.1 (2018-12-10)
 ....................